--- andhow-junit5-extensions/pom.xml
@@ -1,96 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>org.yarnandtail</groupId>
-		<artifactId>andhow-parent</artifactId>
-		<version>1.5.1-SNAPSHOT</version>
-	</parent>
-
-	<artifactId>andhow-junit5-extensions</artifactId>
-
-	<packaging>jar</packaging>
-	<name>AndHow JUnit 5 Extensions</name>
-	<description>
-		Extensions for JUnit 5 to make testing easier for applications using the AndHow framework.
-	</description>
-
-	<dependencies>
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>andhow-shared-test-utils</artifactId>
-			<version>${project.version}</version>
-			<scope>compile</scope>
-		</dependency>
-
-		<!--
-		This package is designed to be used unit testing utils for apps,
-		so dependencies that would normally be test scope are compile scope.
-		However, they are optional so that the downstream app can choose if
-		it wants to use Junit 4 or 5 (and use newer versions than what is
-		used here).
-		-->
-		<dependency>
-			<groupId>org.junit.jupiter</groupId>
-			<artifactId>junit-jupiter-api</artifactId>
-			<scope>compile</scope>
-			<optional>true</optional>
-		</dependency>
-		<dependency>
-			<groupId>org.junit.jupiter</groupId>
-			<artifactId>junit-jupiter-engine</artifactId>
-			<scope>compile</scope>
-			<optional>true</optional>
-		</dependency>
-		<dependency>
-			<groupId>com.github.h-thurow</groupId>
-			<artifactId>simple-jndi</artifactId>
-			<scope>compile</scope>
-			<optional>true</optional>
-		</dependency>
-		<dependency>
-			<!-- Optional dependency of simple-jndi -->
-			<groupId>ch.qos.logback</groupId>
-			<artifactId>logback-classic</artifactId>
-			<optional>true</optional>
-		</dependency>
-		<!-- More traditional test dependencies -->
-		<dependency>
-			<!--
-			A minimal stub implementation that can be tested against, providing
-			clases that exist with types that are expected.
-			-->
-			<groupId>org.yarnandtail</groupId>
-			<artifactId>andhow-test-stubs</artifactId>
-			<version>${project.version}</version>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
-
-	<profiles>
-		<profile>
-			<id>source-and-javadoc-jar</id>
-			<build>
-				<plugins>
-					<plugin>
-						<groupId>org.apache.maven.plugins</groupId>
-						<artifactId>maven-javadoc-plugin</artifactId>
-						<executions>
-							<execution>
-								<id>javadoc-jar</id>
-								<phase>package</phase>
-								<goals>
-									<goal>jar</goal>
-								</goals>
-								<configuration>
-									<includeDependencySources>false</includeDependencySources>
-								</configuration>
-							</execution>
-						</executions>
-					</plugin>
-				</plugins>
-			</build>
-		</profile>
-	</profiles>
-
-</project>
