+++ junit5-extensions/junit5-extensions-no-andhow-dependency/src/main/java/org/yarnandtail/andhow/junit5/ext/ExtensionBase.java
