+++ pom.xml
@@ -87,8 +87,10 @@
 		<module>andhow-testing/andhow-simulated-app-tests</module>
 		<module>andhow-testing/andhow-annotation-processor-test-harness</module>
 		<module>andhow-shared-test-utils</module>
-		<module>andhow-junit5-extensions</module>
 		<module>andhow-test-stubs</module>
+		<module>junit5-extensions/andhow-junit5-extensions</module>
+		<module>junit5-extensions/junit5-extensions-no-andhow-dependency</module>
+		<module>junit5-extensions/junit5-extensions-with-andhow-dependency</module>
 	</modules>
 
 	<repositories>
@@ -266,6 +268,16 @@
 					<artifactId>maven-enforcer-plugin</artifactId>
 					<version>3.0.0</version>
 				</plugin>
+				<plugin>
+					<groupId>org.apache.maven.plugins</groupId>
+					<artifactId>maven-surefire-plugin</artifactId>
+					<version>3.0.0-M7</version>
+				</plugin>
+				<plugin>
+					<groupId>org.apache.maven.plugins</groupId>
+					<artifactId>maven-failsafe-plugin</artifactId>
+					<version>3.0.0-M7</version>
+				</plugin>
 			</plugins>
 		</pluginManagement>
 		<plugins>
@@ -296,12 +308,10 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-surefire-plugin</artifactId>
-				<version>3.0.0-M7</version>
 			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-failsafe-plugin</artifactId>
-				<version>3.0.0-M7</version>
 				<executions>
 					<execution>
 						<goals>
